feat(layout)!: client layout component
